refactor: :recycle: move splash screen images to SplashScreen class
